Refactor on edit student modal in preparation for adding the additional feed view.
