+++ client/src/pages/Dashboard/views/Students/index.js
@@ -6,7 +6,7 @@ import { useTopbarConfig } from "pages/Dashboard/components/Topbar";
 
 import StudentList from "./components/StudentList";
 import { ModalProvider } from "components/Modal";
-import EditStudentModal from "../../components/EditStudentModal"
+import StudentModal from "../../components/StudentModal"
 import StudentListControls from "./components/StudentListControls";
 
 const Students = () => {
@@ -21,7 +21,7 @@ const Students = () => {
             <ModalProvider>
                 <StudentListControls sort={[ sort, setSort ]} groupBy={[ groupBy, setGroupBy ]} search={[ search, setSearch ]} />
                 <StudentList sort={sort} groupBy={groupBy} search={search.toLowerCase()} />
-                <EditStudentModal />
+                <StudentModal />
             </ModalProvider>
         </Section>
     );
